test: separte common methods
